requirements.txt
python-decouple
psycopg2-binary
django
Pillow
djangorestframework